requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
alpaca==1.0.0
alpaca-py==0.13.6
annotated-types==0.6.0
attrs==23.2.0
cattrs==23.2.3
certifi==2023.11.17
charset-normalizer==3.3.2
deprecation==2.1.0
exchange_calendars==4.5.1
flatbuffers==23.5.26
frozenlist==1.4.1
idna==3.6
korean-lunar-calendar==0.3.1
msgpack==1.0.3
multidict==6.0.4
numpy==1.26.3
openmeteo_requests==1.1.0
openmeteo_sdk==1.7.1
packaging==23.2
pandas==1.5.3
pandas_market_calendars==4.3.3
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
pyluach==2.2.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-cache==1.1.1
retry-requests==2.0.0
six==1.16.0
sseclient-py==1.8.0
toolz==0.12.0
typing_extensions==4.9.0
tzdata==2023.4
url-normalize==1.4.3
urllib3==1.26.18
websocket-client==1.7.0
websockets==11.0.3
yarl==1.9.4
